Move validate_circuit_data to models layer, fix grading dependency
Extract validate_circuit_data from controllers/file_controller.py into
models/circuit_validator.py as its canonical location. The grading layer
now imports from models instead of controllers, eliminating the backwards
dependency. Controllers re-export for backward compatibility.

Closes #588
